// generated/clients/javascript/services/uigf-game-global-communityrecords/src/api/genshinChallengeApi/options.ts
// Copyright (c) Microsoft Corporation.
// Licensed under the MIT License.

import { OperationOptions } from "@azure-rest/core-client";

/** Optional parameters. */
export interface GenshinChallengeApiGetStygianOnslaughtOptionalParams extends OperationOptions {
  scheduleType?: number;
  needDetail?: boolean;
  needAll?: boolean;
}

/** Optional parameters. */
export interface GenshinChallengeApiGetImaginariumTheaterOptionalParams extends OperationOptions {
  scheduleType?: number;
  needDetail?: boolean;
  needAll?: boolean;
}

/** Optional parameters. */
export interface GenshinChallengeApiGetSpiralAbyssOptionalParams extends OperationOptions {
  scheduleType?: number;
  needDetail?: boolean;
  needAll?: boolean;
}